Rename `TableAsset` -> `TableEntry`
