+++ web-common/src/features/tables/TableEntry.svelte
